I have changed several things like the memory mngr is now a static class that keeps track of the memory. I also made several classes that make the app and the game run. ConsoleApp and BaseGame. You need to inherit from baseclass and override several member functions and pass it in to the ConsoleApp to run your game. Next up is to set up the render system to not be called directly. we will call a system that will queue commands to send to the renderer.
