+++ DoomConsoleGame/ConsoleDoomGame/ConsoleDoomGame.vcxproj
@@ -140,6 +140,10 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="Src\Demo.cpp" />
+    <ClCompile Include="Src\DemoGame.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="Src\DemoGame.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
