+++ DoomConsoleGame/ConsoleDoomGame/ConsoleDoomGame.vcxproj.filters
@@ -10,5 +10,13 @@
     <ClCompile Include="Src\Demo.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="Src\DemoGame.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="Src\DemoGame.h">
+      <Filter>Source Files</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
