+++ DoomConsoleGame/GameEngine/Src/ConsoleRenderer.h
@@ -12,20 +12,28 @@ render the current game
 #ifndef CRenderer_H
 #define CRenderer_H
 
+#include "platform.h"
+
+#pragma pack(push, 16)
 struct ScreenBuffer
 {
 	int Width;
 	int Height;
+	wchar_t *Buffer;
+	WORD	*Attribs;
+	float	*DepthBuffer;
+	HANDLE	consoleHandle;
 };
+#pragma pack(pop)
 
-int screenWidth = 640;// 320; //160;
-int screenHeight = 176; //88;
-wchar_t* screenBuffer = 0;
-unsigned short* screenAttribs = 0;
-float *depthBuffer = 0;
-
-int fontw = 4;// 8;
-int fonth = 4;// 8;
+//int screenWidth = 640;// 320; //160;
+//int screenHeight = 176; //88;
+//wchar_t* screenBuffer = 0;
+//unsigned short* screenAttribs = 0;
+//float *depthBuffer = 0;
+//
+//int fontw = 4;// 8;
+//int fonth = 4;// 8;
 
 // Types of Pixels to use
 #define PIXEL_SOLID 0x2588
@@ -51,65 +59,88 @@ int fonth = 4;// 8;
 #define PIXEL_COLOR_GREY			PIXEL_COLOR_DARK_RED | PIXEL_COLOR_DARK_GREEN | PIXEL_COLOR_DARK_BLUE
 #define PIXEL_COLOR_WHITE			PIXEL_COLOR_GREY | PIXEL_COLOR_INTENSITY
 
-// Draws a single pixel onto the screen buffer using the desired x, y position, type of pixel and color.
-void DrawPixel(int x, int y, wchar_t pixel = PIXEL_SOLID, unsigned short color = PIXEL_COLOR_WHITE)
+class ConsoleRenderer
 {
-	if (x < 0)
+public:
+	ConsoleRenderer() {}
+	~ConsoleRenderer() {}
+	
+	ScreenBuffer* GetRenderBuffers() { return screen; }
+
+	void SetScreenBuffer(ScreenBuffer *_screen)
 	{
-		x = 0;
-	}
-	if (x >= screenWidth)
-	{
-		x = screenWidth - 1;
-	}
-	if (y < 0)
-	{
-		y = 0;
-	}
-	if (y >= screenHeight)
-	{
-		y = screenHeight - 1;
+		screen = _screen;
+		Assert(screen);
+		// Check to see if we were able to allocate memory from our pool.
+		Assert(screen->Buffer);
+		Assert(screen->Attribs);
+		Assert(screen->DepthBuffer);
 	}
 
-	screenBuffer[y * screenWidth + x] = pixel;
-	screenAttribs[y * screenWidth + x] = color;
-}
-
-// Draws a sprite onto the screen buffer using desired x, y pos and color.
-void DrawSprite(int posX, int posY, int spriteW, int spriteH, wchar_t* spriteData, unsigned short *colorData)
-{
-	for (int col = 0; col < spriteH; ++col)
+	// Draws a single pixel onto the screen buffer using the desired x, y position, type of pixel and color.
+	void DrawPixel(int x, int y, wchar_t pixel = PIXEL_SOLID, unsigned short color = PIXEL_COLOR_WHITE)
 	{
-		for (int row = 0; row < spriteW; ++row)
+		if (x < 0)
 		{
-			int pixelX = posX + row;
-			int pixelY = posY + col;
-			wchar_t glyph = spriteData[col * spriteW + row];
-			unsigned short color = colorData[col * spriteW + row];
-			DrawPixel(pixelX, pixelY, glyph, color);
+			x = 0;
+		}
+		if (x >= screen->Width)
+		{
+			x = screen->Width - 1;
+		}
+		if (y < 0)
+		{
+			y = 0;
+		}
+		if (y >= screen->Height)
+		{
+			y = screen->Height - 1;
+		}
+
+		screen->Buffer[y * screen->Width + x] = pixel;
+		screen->Attribs[y * screen->Width + x] = color;
+	}
+
+	// Draws a sprite onto the screen buffer using desired x, y pos and color.
+	void DrawSprite(int posX, int posY, int spriteW, int spriteH, wchar_t* spriteData, unsigned short *colorData)
+	{
+		for (int col = 0; col < spriteH; ++col)
+		{
+			for (int row = 0; row < spriteW; ++row)
+			{
+				int pixelX = posX + row;
+				int pixelY = posY + col;
+				wchar_t glyph = spriteData[col * spriteW + row];
+				unsigned short color = colorData[col * spriteW + row];
+				DrawPixel(pixelX, pixelY, glyph, color);
+			}
 		}
 	}
-}
 
-// Clears the entire screen buffer to the desired color.
-void ClearBuffer(short color = PIXEL_COLOR_BLACK)
-{
-	for (int y = 0; y < screenHeight; ++y)
+	// Clears the entire screen buffer to the desired color.
+	void ClearBuffer(short color = PIXEL_COLOR_BLACK)
 	{
-		for (int x = 0; x < screenWidth; ++x)
+		for (int y = 0; y < screen->Height; ++y)
 		{
-			DrawPixel(x, y, PIXEL_SOLID, color);
+			for (int x = 0; x < screen->Width; ++x)
+			{
+				DrawPixel(x, y, PIXEL_SOLID, color);
+			}
 		}
 	}
-}
 
-// Draws the entire buffer to the screen
-void PresentBuffer(HANDLE handle)
-{
-	DWORD bytesWritten = 0;
-	screenBuffer[screenWidth * screenHeight - 1] = '\0';
-	WriteConsoleOutputCharacter(handle, screenBuffer, screenWidth * screenHeight, { 0, 0 }, &bytesWritten);
-	WriteConsoleOutputAttribute(handle, screenAttribs, screenWidth * screenHeight, { 0, 0 }, &bytesWritten);
-}
+	// Draws the entire buffer to the screen
+	void PresentBuffer()
+	{
+		DWORD bytesWritten = 0;
+		screen->Buffer[screen->Width * screen->Height - 1] = '\0';
+		WriteConsoleOutputCharacter(screen->consoleHandle, screen->Buffer, screen->Width * screen->Height, { 0, 0 }, &bytesWritten);
+		WriteConsoleOutputAttribute(screen->consoleHandle, screen->Attribs, screen->Width * screen->Height, { 0, 0 }, &bytesWritten);
+	}
+
+protected:
+	ScreenBuffer *screen;
+
+};
 
 #endif
